+++ examples.py
@@ -1,21 +1,10 @@
 import os
-import sys
 
-from loguru import logger
 from AzApi.AzApi import AzApi
-from dotenv import load_dotenv
 import tempfile
 
 from AzApi.utils.AzApi_agents import AgentsBy
 from AzApi.utils.AzApi_boards import WorkItemsDef, WorkItemsStatesDef
-
-logger.remove()
-logger.add(
-    sys.stderr,
-    level="TRACE",
-    format="<green>{time:YYYY-MM-DD HH:mm:ss.SSS}</green> | <level>{level: <8}</level> | <cyan>{name}</cyan>:<cyan>{function}</cyan>:<cyan>{line}</cyan> - <level>{message}</level>",
-)
-load_dotenv()
 
 # Examples:
 ## Init
@@ -48,18 +37,18 @@ api.Boards.change_work_item_state(testcase_id, WorkItemsStatesDef.TestCase.Close
 
 # Repo Examples:
 ## Clone repository without history
-api.Repo.clone_repository(tempfile.gettempdir(), branch="main", depth=1, submodules=True)
+api.Repos.clone_repository(tempfile.gettempdir(), branch="main", depth=1, submodules=True)
 
 ## Get list of all available branches
-all_branches_list = api.Repo.get_all_branches()
+all_branches_list = api.Repos.get_all_branches()
 
 ## Check available pull requests
-prs = api.Repo.get_active_pull_requests()
+prs = api.Repos.get_active_pull_requests()
 
 ## Create Pull Request
-pr_id = api.Repo.create_pr("Test PullRequest", "TestBranch", "main", "Testing API Request.")
+pr_id = api.Repos.create_pr("Test PullRequest", "TestBranch", "main", "Testing API Request.")
 ## Add Reviewer to PR
-api.Repo.add_pr_reviewer(pr_id, "user1@gmail.com")
+api.Repos.add_pr_reviewer(pr_id, "user1@gmail.com")
 
 # Agents Examples:
 ## Add Agents userCapabilies